refactor(dish): reorganize files
