+++ src/models/morning/dishes/Toast.ts
@@ -1,12 +1,7 @@
 import { IDish } from '@/interfaces/bakery/IDish';
-import { NonRepeatableDish } from '@/interfaces/bakery/NonRepeatableDish';
+import { NonRepeatableDish } from '@/models/NonRepeatableDish';
 
 export class Toast extends NonRepeatableDish implements IDish {
   public type = 'side';
   public name = 'toast';
-  public have_seconds_times = 1;
-
-  public print(): string {
-    return this.have_seconds_times > 1 ? `${this.name}(${this.have_seconds_times}x)` : `${this.name}`;
-  }
 }
